fix seed file, first store data to json because heroku has conflict with open-uri
